refactor(app): extract startup loops to meow_app lib modules
Move `subscription_refresh_loop` and the `geodata_fetch_missing_on_startup`
+ `geodata_auto_update_loop` helpers out of `main.rs` and into dedicated
modules under `meow-app/src/`. Downstream FFI callers that build a
`Tunnel` directly can now wire the same auto-refresh + geodata startup
behavior without duplicating the loop bodies.

- New `meow_app::subscription_refresh::run_loop`
- New `meow_app::geodata_fetch::{run_on_startup, auto_update_loop}`
- `main.rs` shrinks from 815 → 606 lines; call sites updated
- New integration test `geodata_fetch_test.rs` covers the extracted helpers
- New `config_smoke.rs` cases exercise the verbatim community Clash Meta
  fixture so parser gaps surface as test failures, not silent ignores
